requirements.txt
# Core
anthropic>=0.25
python-dotenv>=1.0
httpx>=0.27

# YouTube API
google-api-python-client>=2.100
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2

# Google Sheets
gspread>=6.0
pandas>=2.0

# Video processing
torch>=2.0          # Silero VAD
torchaudio>=2.0

# Transcription
openai-whisper>=20230918

# Optional
# replicate>=0.25   # AI thumbnail generation
# Pillow>=10.0      # Image processing
